refactor: Fix lint errors and make some modules files instead of folders
