+++ utilities/ideascale-importer/ideascale_importer/app.py
@@ -1,10 +1,10 @@
+"""CLI entrypoint."""
+
 import typer
-import ideascale_importer.cli.db
 import ideascale_importer.cli.ideascale
 import ideascale_importer.cli.snapshot
 
 app = typer.Typer(add_completion=False)
-app.add_typer(ideascale_importer.cli.db.app, name="db", help="Postgres DB commands (e.g. seeding)")
 app.add_typer(ideascale_importer.cli.ideascale.app, name="ideascale", help="IdeaScale commands (e.g. importing data)")
 app.add_typer(ideascale_importer.cli.snapshot.app, name="snapshot", help="Snapshot commands (e.g. importing data)")
