+++ utilities/ideascale-importer/ideascale_importer/ideascale/__init__.py
@@ -1,0 +1,1 @@
+"""IdeaScale importer module."""
