+++ utilities/ideascale-importer/pyproject.toml
@@ -54,8 +54,9 @@ target-version = ['py311']
 [tool.ruff]
 line-length = 132
 target-version = 'py311'
-# add pydoc lints
+exclude = ["tests"]
 extend-select = ["D"]
+
 [build-system]
 requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
